+++ packages/sdks/src/components/content/content.lite.tsx
@@ -181,6 +181,7 @@ export default function ContentComponent(props: ContentProps) {
 
   return (
     <EnableEditor
+      apiHost={props.apiHost}
       nonce={props.nonce}
       content={props.content}
       data={props.data}